Refactor App to Singleton pattern
