+++ yarte_codegen/src/wasm/client/tests/mod.rs
@@ -70,7 +70,7 @@ fn test() {
 
             #[doc(hidden)]
             #[inline]
-            fn __render (&mut self, __addr: &yarte_wasm_app::Addr<Self>) {
+            fn __render (&mut self, __addr: &'static yarte_wasm_app::Addr<Self>) {
                 if self.black_box.t_root == <u8 as yarte_wasm_app::YNumber>::zero() {
                     return;
                 }
@@ -84,10 +84,10 @@ fn test() {
 
             #[doc(hidden)]
             #[inline]
-            fn __hydrate (&mut self, __addr: &yarte_wasm_app::Addr<Self>) { }
+            fn __hydrate (&mut self, __addr: &'static yarte_wasm_app::Addr<Self>) { }
 
             #[doc(hidden)]
-            fn __dispatch(&mut self, __msg: Self::Message, __addr: &yarte_wasm_app::Addr<Self>) {
+            fn __dispatch(&mut self, __msg: Self::Message, __addr: &'static yarte_wasm_app::Addr<Self>) {
                 use Msg::*;
                 match __msg {
                     Foo => foo(self, __addr)
