+++ yarte_codegen/src/wasm/client/tests/tree_diff.rs
@@ -34,7 +34,7 @@ fn test_diff_u16() {
 
             #[doc(hidden)]
             #[inline]
-            fn __render (&mut self, __addr: &yarte_wasm_app::Addr<Self>) {
+            fn __render (&mut self, __addr: &'static yarte_wasm_app::Addr<Self>) {
                 if self.black_box.t_root == <u16 as yarte_wasm_app::YNumber>::zero() {
                     return;
                 }
@@ -48,10 +48,10 @@ fn test_diff_u16() {
 
             #[doc(hidden)]
             #[inline]
-            fn __hydrate (&mut self, __addr: &yarte_wasm_app::Addr<Self>) { }
+            fn __hydrate (&mut self, __addr: &'static yarte_wasm_app::Addr<Self>) { }
 
             #[doc(hidden)]
-            fn __dispatch(&mut self, __msg: Self::Message, __addr: &yarte_wasm_app::Addr<Self>) {
+            fn __dispatch(&mut self, __msg: Self::Message, __addr: &'static yarte_wasm_app::Addr<Self>) {
                 use Msg::*;
                 match __msg {
                     Foo => foo(self, __addr)
@@ -129,7 +129,7 @@ fn test_diff_u16_1() {
 
             #[doc(hidden)]
             #[inline]
-            fn __render (&mut self, __addr: &yarte_wasm_app::Addr<Self>) {
+            fn __render (&mut self, __addr: &'static yarte_wasm_app::Addr<Self>) {
                 if self.black_box.t_root == <u16 as yarte_wasm_app::YNumber>::zero() {
                     return;
                 }
@@ -147,10 +147,10 @@ fn test_diff_u16_1() {
 
             #[doc(hidden)]
             #[inline]
-            fn __hydrate (&mut self, __addr: &yarte_wasm_app::Addr<Self>) { }
+            fn __hydrate (&mut self, __addr: &'static yarte_wasm_app::Addr<Self>) { }
 
             #[doc(hidden)]
-            fn __dispatch(&mut self, __msg: Self::Message, __addr: &yarte_wasm_app::Addr<Self>) {
+            fn __dispatch(&mut self, __msg: Self::Message, __addr: &'static yarte_wasm_app::Addr<Self>) {
                 use Msg::*;
                 match __msg {
                     Foo => foo(self, __addr)
