+++ yarte_wasm_app/benches/src/lib.rs
@@ -3,7 +3,7 @@ extern crate serde_derive;
 
 use wasm_bindgen::prelude::*;
 
-use yarte_wasm_app::App;
+use yarte_wasm_app::run;
 
 use crate::app::NonKeyed;
 
@@ -14,7 +14,7 @@ mod row;
 
 #[wasm_bindgen(start)]
 pub fn start() {
-    let app = NonKeyed::start_default();
+    let app = run!(NonKeyed);
     unsafe {
         app.hydrate();
     }
